# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HSE Mobile Application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        picGroupId:
          type: integer
        picGroupName:
          type: string
        color:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - riskLevel
        - createdAt
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        picGroupId:
          type: integer
        color:
          type: string
      required:
        - name
        - riskLevel
...